remove single-use set-workspaces mixin
